Change when/how boosts are applied
This is a major change.
Boosts are now applied (safely) at EnterLevel and EnterCombat as needed.
Precomputation occurs at LevelStarted, and everything can be recomputed
on PingRequested.
The net effect of this is enemies will have the "right" health and stats
when you first see them and enemies joining in combat will also compute
correctly.
Since this is all guarded by a LCC_PASSIVE passive, boosts will only be
applied once.

Most things are now added using boosts to simplify removal (as boosts
can be tracked to a Cause).
   This includes all normal boosts, actions, abilities, special
   resources, and now spells.
Only Passives are added seperately as there is no way to add them via
boost, and they cannot be tracked to a Cause.
As such, Passives are added withing a LCC_PASSIVE_START and
LCC_PASSIVE_END passive block. This allows (in most cases except after
fleeing ... more on that later[1]), removal by slicing passives with these
indexes, and removing the ones in between.

Now boosts and passives are only removed when a ping is requested (new
config)[1]

Removed the LCC_* statuses, since the passives work just as well.

[1] A special case is handled for combat ended. Passives are removed
because after rejoining combat, for some reason, all passives are
reshuffled and we lose the LCC_PASSIVE_START and LCC_PASSIVE_END
indexers. For this we use a series of extra passives
LCC_PASSIVE_PASSIVED (and LCC_PASSIVE_BOOSTED) to check if we have
removed partial boosts